Handle errors in the repository. Create DatabaseError type
