+++ src/config/index.ts
@@ -1,11 +1,28 @@
 import { z } from 'zod';
 import dotenv from 'dotenv';
 import { getCommitHash } from '../utils/utils';
-import { ConfigValidationSchema } from '../validation-schema';
+// import { ConfigValidationSchema } from '../validation-schema';
 
 dotenv.config();
 export const DBConfigValidationSchema = z.object({
   url: z.string().url()
+});
+
+const PORT_REGEX = /^(?:[1-9]\d{0,3}|[1-5]\d{4}|6[0-4]\d{3}|65[0-4]\d{2}|655[0-2]\d|6553[0-5])$/;
+const validatePort = (port: string) => {
+  if (!PORT_REGEX.test(port)) {
+    throw new Error('Invalid PORT number');
+  }
+
+  return Number(port);
+};
+
+export const ConfigValidationSchema = z.object({
+  appEnv: z.string(),
+  dbConfig: z.string().transform((val) => DBConfigValidationSchema.safeParse(JSON.parse(val))),
+  serverPort: z.string().transform((port) => validatePort(port)),
+  lastCommitHash: z.string(),
+  logLevel: z.string()
 });
 
 export enum AppEnv {
